Keep the preset catalogue out of the landing page's bundle
The landing page was a client component importing the whole preset catalogue -
686 lines, 57 entries with descriptions and prompts - to render a count and six
cards that use four fields each. It now has a server shell that reads the
catalogue and passes down only what is rendered, which is the pattern Next
intends and keeps the table on the server.

The ornament test was matching on a filename, so moving that markup would have
turned the check off silently. It now finds the markup by content.

A note on what this does and does not buy, since the numbers are easy to
misread. Locally the landing page measured 78 before and 87 after, but repeat
runs of the same build swung LCP between 3.5s and 4.4s: Chrome and the Next
server competing for this machine. Against production the page scores 93 and 98
on consecutive runs, so the local figure was mostly environmental and the honest
verification here is that the import is gone, not that a score moved.

The preset data still arrives on the landing page, because Next prefetches the
routes it links to at low priority. That is deliberate and makes navigation
instant; disabling it would trade real behaviour for a metric. /about already
reads the catalogue on the server, and the gallery, create and editor pages each
genuinely need what they import.
